Implement a consistent request management
Important commit which needed an extensive testing and a massive work
load. Requests are now handled in a consistent way. Filters are executed
in order, ExtensionConcealment being the last. This filter adds a .jsp
extension if the request is for a jsp page which isn't
associated with any servlet.
This allow for all extensions and to masked and for a consistent naming.
Resources are now named in lower-case, separared by dash. Login/ became
login/, and so on. As a result, there is not anymore the problem of being
able to access directly .jsp files which are supposed to be used by
servlets.